Data Structure Simulator
Data Structure simulation <1.0.0> (DSS) is an application where all the
data structures with all its functionality are graphically demonstrated
. The application will also contain descriptions of all the data
structures in two different languages namely Hindi and English. The
codes in the description are also written in two technical languages
namely C and JAVA. The application’s objective is to teach all the
students how different data structures works and to give them a crystal
clear concept of its working. It can also be used by teachers as manual
for teaching purpose.